+++ frostmourne-spi-starter/src/main/java/com/autohome/frostmourne/spi/starter/config/FrostmourneSpiProperties.java
@@ -7,21 +7,11 @@ public class FrostmourneSpiProperties {
 
     private String serviceAddr;
 
-    private Boolean mock;
-
     public String getServiceAddr() {
         return serviceAddr;
     }
 
     public void setServiceAddr(String serviceAddr) {
         this.serviceAddr = serviceAddr;
-    }
-
-    public Boolean getMock() {
-        return mock;
-    }
-
-    public void setMock(Boolean mock) {
-        this.mock = mock;
     }
 }